changed castling_rights and en_passant_square to 8 byte int

+++ include/position.hpp
@@ -32,10 +32,10 @@ struct position {
 	int8_t side_to_move;
 
 	/* castling rights indexed by piece color.                               */
-	int castling_rights[2];
+	int8_t castling_rights[2];
 
 	/* en passant square, may be `NO_SQUARE`.                                */
-	int en_passant_square;
+	int8_t en_passant_square;
 };
 
 /* print out information about the position. useful for debugging.           */
